tidy to non tidy conversion: Lots of changes here.
1. Removing eco_demo and eco_run function, will eventually bring eco_run back. Removed because the internal data which these functions run off of has changed.

2. Internal data has been converted from "tidy" data to "non tidy" data. Reason being that tidy data was growing the dataset unnecessarily large (and early on). Now, data is "non tidy" until the very end. This has increased the speed on larger dataset quite a bit. What took 2 and half minutes now takes 15 seconds.

3. Eco run all now does all the work on unique dbh/common_name pairs to avoid duplicating work. At the end, the unique results are joined to the entire inventory with all it's duplicate records.

4. As mentioned, the internal data has been changed dramatically, names have changes, columns have changed, and the structure has changed.

All of this was done for speed. While it's gotten a lot faster, speed is currently dependent on the unique dbh/common_name pairs because we do grow the dataset x15 for each record when joining to the benefits. This is the only "tidy" column that has a lot of the work done to it. I haven't figured out how to keep this untidy.