Couple things:
Added warning to eco_guess when x is factor and coerces to character. Previously a message but warning seems more appropriate.

Removed some of the global var issues in eco_run, this removes a few instances of that note which needs to eventually be resolved, not the end of the world for now.

eco_guess now has n as an additional argument for setting the guess threshold, defaults to .8 if nothing is given. Warnings have also been added if n > 1 or < 0, spits out a warning and defaults to 0.8

extract_matches, added the n arg for use in eco_guess as extract_matches is a utility funciton for eco_run_all

Added tests for the stuff mentioned above.

Updated the docs and NEWS.md to reflect these changes